pyproject.toml
[project]
name = "mnexa"
version = "0.0.1"
description = "A disciplined wiki maintainer for a personal markdown knowledge base."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.scripts]
mnexa = "mnexa.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mnexa"]

[tool.hatch.build.targets.wheel.force-include]
"src/mnexa/templates/CLAUDE.md" = "mnexa/templates/CLAUDE.md"
"src/mnexa/templates/index.md" = "mnexa/templates/index.md"
"src/mnexa/templates/gitignore" = "mnexa/templates/gitignore"
"src/mnexa/prompts/stage1.md" = "mnexa/prompts/stage1.md"
"src/mnexa/prompts/stage2.md" = "mnexa/prompts/stage2.md"
"src/mnexa/prompts/query.md" = "mnexa/prompts/query.md"
"src/mnexa/prompts/lint.md" = "mnexa/prompts/lint.md"

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
pdf = ["pdfplumber>=0.11"]
docx = ["python-docx>=1.1"]

[dependency-groups]
dev = [
    "pyright>=1.1.380",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "types-pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "PTH",  # use pathlib
    "TID",  # tidy imports
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Argument", "typer.Option"]

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = "warning"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"
